jupyterhub-docker/middleware/pyproject.toml
[project]
name = "middleware"
version = "0.1.0"
description = "middleware uv project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-dotenv",
    "rapidfuzz",
    "thefuzz",        
    "fluent-logger",
    "httpx"  
]
